set CRAN mirrors option to reflect user prefs
