+++ src/cpp/session/CMakeLists.txt
@@ -104,6 +104,7 @@ set (SESSION_SOURCE_FILES
    modules/SessionBreakpoints.cpp
    modules/SessionCodeSearch.cpp
    modules/SessionConfigFile.cpp
+   modules/SessionCRANMirrors.cpp
    modules/SessionCrashHandler.cpp
    modules/SessionUserCommands.cpp
    modules/SessionConsole.cpp
